Refactored test_ship_sunk_submarine_player into test_submarine.
